minor refactoring according to best practices
