+++ compressors/gzip/gzip_compressor.go
@@ -1,24 +1,26 @@
-package supportedcompressors
+package gzip
 
 import (
 	"bytes"
 	"compress/gzip"
 	"fmt"
 	"io/ioutil"
+
+	"github.com/open-cluster-management/hub-of-hubs-message-compression/compressors"
 )
-
-// CreateGzipCompressor returns a new instance of gzip-based compressor.
-func CreateGzipCompressor() *GZipCompressor {
-	return &GZipCompressor{}
-}
-
-// GZipCompressor implements Compressor with gzip-based logic.
-type GZipCompressor struct{}
 
 const gzipCompressorErrorString = "gzip compressor error"
 
+// NewGzipCompressor returns a new instance of gzip-based compressor.
+func NewGzipCompressor() compressors.Compressor {
+	return &CompressorGZip{}
+}
+
+// CompressorGZip implements Compressor with gzip-based logic.
+type CompressorGZip struct{}
+
 // Compress compresses a slice of bytes using gzip lib.
-func (gzc *GZipCompressor) Compress(data []byte) ([]byte, error) {
+func (gzc *CompressorGZip) Compress(data []byte) ([]byte, error) {
 	var buf bytes.Buffer
 
 	gw, err := gzip.NewWriterLevel(&buf, gzip.BestSpeed)
@@ -34,9 +36,9 @@ func (gzc *GZipCompressor) Compress(data []byte) ([]byte, error) {
 }
 
 // Decompress decompresses a slice of gzip-compressed bytes using gzip lib.
-func (gzc *GZipCompressor) Decompress(compressed []byte) ([]byte, error) {
-	// Write gzipped data to the client
-	gr, err := gzip.NewReader(bytes.NewBuffer(compressed))
+func (gzc *CompressorGZip) Decompress(compressedData []byte) ([]byte, error) {
+	// create a reader for the gzipped data
+	gr, err := gzip.NewReader(bytes.NewBuffer(compressedData))
 	if err != nil {
 		return nil, fmt.Errorf("%s - %w", gzipCompressorErrorString, err)
 	}
